requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
deepl==1.18.0
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
faiss-cpu==1.8.0.post1
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.25
langchain-openai==0.1.19
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langsmith==0.1.94
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.0
numpy==1.26.4
openai==1.37.1
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pillow==10.2.0
platformdirs==4.2.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.10
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.12
tbb==2021.11.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.40.2
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.26.1
watchfiles==0.21.0
websockets==12.0
wheel==0.43.0
yarl==1.9.4
